@@ -1,8 +1,13 @@
-import * as React from "react";
-import ReactDOM from "react-dom";
-import { App } from "./App";
+import React from "react";
+import ReactDOM from "react-dom/client";
+import "./index.css";
+import App from "./pages/App";
 
-ReactDOM.render(
-    <App/>,
-    document.getElementById("root")
-)
+const root = ReactDOM.createRoot(
+  document.getElementById("root") as HTMLElement
+);
+root.render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>
+);
