@@ -1,5 +1,5 @@
 
-export interface GameState {
+export type GameState = {
     players: [ Player, Player ]; // a player array contains exactly two Players
     gameStatus: {
         endOfGame: boolean;
@@ -7,14 +7,14 @@ export interface GameState {
     };
 }
 
-interface Player {
+export type Player = {
     name: string;
     pits: Pit[];
     type: "player1" | "player2"; // only "player1" and "player2" are valid options for this string
     hasTurn: boolean;
 }
 
-interface Pit {
+export type Pit = {
     index: number;
     nrOfStones: number;
 }
